Some refactor to avoid duplication
